requirements.txt
absl-py==0.9.0
astor==0.8.1
boto3==1.12.20
botocore==1.15.20
certifi==2020.11.8
chardet==3.0.4
cloudpickle==1.4.1
cycler==0.10.0
Cython==0.29.15
cytoolz==0.10.1
dask==2.16.0
decorator==4.4.2
docutils==0.15.2
easydict==1.9
gast==0.2.2
google-pasta==0.1.8
grpcio==1.27.2
h5py==2.10.0
idna==2.9
imageio==2.8.0
jmespath==0.9.5
json-lines==0.5.0
jsonlines==1.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
lmdb==0.94
Markdown==3.2.1
matplotlib==3.2.0
msgpack==1.0.0
msgpack-numpy==0.4.4.3
networkx==2.4
numpy==1.18.1
olefile==0.46
opt-einsum==3.2.0
Pillow==7.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-transformers==1.1.0
PyWavelets==1.1.1
PyYAML==5.1.2
pyzmq==19.0.0
regex==2020.2.20
requests==2.23.0
s3transfer==0.3.3
scikit-image==0.16.2
scipy==1.4.1
sentencepiece==0.1.85
six==1.14.0
tabulate==0.8.6
tensorboard==1.15.0
tensorboardX==1.8
tensorflow==1.15.2
tensorflow-estimator==1.15.1
tensorpack==0.9.4
termcolor==1.1.0
toolz==0.10.0
torch==1.4.0
torchvision==0.5.0
tornado==6.0.4
tqdm==4.31.1
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.12.1
